HID: Avoid ABI change in 4.17.6
Commit 8f732850df1b "HID: core: allow concurrent registration of
drivers" introduced atomic bit-operations on hid_device::status, and
changed its type from unsigned int to unsigned long as required for
those operations.

Revert the type change and use cmpxchg() for the bit-operations,
since it supports unsigned int.


Gbp-Pq: Topic debian
Gbp-Pq: Name hid-avoid-abi-change-in-4.17.6.patch

+++ include/linux/hid.h
@@ -569,7 +569,7 @@ struct hid_device {							/* device report descriptor */
 	bool battery_avoid_query;
 #endif
 
-	unsigned long status;						/* see STAT flags above */
+	unsigned int status;						/* see STAT flags above */
 	unsigned claimed;						/* Claimed by hidinput, hiddev? */
 	unsigned quirks;						/* Various quirks the device can pull on us */
 	bool io_started;						/* If IO has started */
